Implement E2E testing for filtering
Implemented the end-to-end tests for filtering todos. Also edited a method name inside of todo-list.po from "selectRole" to "selectStatus" to better describe the method.

Added "Todos" as a possible argument in the getNavLink method inside of app.po in order to test that the page navigates properly with the Todos page.